Extract `classifieds::api` module
